json data reading feature, still need work
