+++ .idea/PharmacyManagementSystem.iml
@@ -7,5 +7,6 @@
     </content>
     <orderEntry type="inheritedJdk" />
     <orderEntry type="sourceFolder" forTests="false" />
+    <orderEntry type="library" name="json-simple-1.1" level="project" />
   </component>
 </module>
