+++ .idea/libraries/json_simple_1_1.xml
@@ -1,0 +1,9 @@
+<component name="libraryTable">
+  <library name="json-simple-1.1">
+    <CLASSES>
+      <root url="jar://$USER_HOME$/Downloads/json-simple-1.1.jar/json-simple-1.1.jar!/" />
+    </CLASSES>
+    <JAVADOC />
+    <SOURCES />
+  </library>
+</component>
