+++ src/main/java/medTracker/patients.json
@@ -1,0 +1,2 @@
+{"phoneNumber":"123-456-7890","name":"John Doe","id":1,"age":30}
+{"phoneNumber":"987-654-3210","name":"Jane Smith","id":2,"age":25}
